kotlin-cli: try out mordant multiplatform 2 beta
fails because not available on macosx

https://scans.gradle.com/s/hlci6h3lc34xc/failure
