+++ versions.properties
@@ -22,9 +22,11 @@ version.com.apollographql.apollo..apollo-runtime=2.5.6
 
 version.com.beust..klaxon=5.5
 
-version.com.github.ajalt..mordant=1.2.1
+version.com.github.ajalt..mordant=2.0.0-beta1
 
 version.com.github.ajalt.clikt..clikt=3.1.0
+
+version.com.github.ajalt.mordant..mordant=2.0.0-beta1
 
 version.com.github.kittinunf.fuel..fuel=2.3.1
